requirements.txt
faiss-cpu==1.13.2
groq==0.37.1
huggingface_hub==1.5.0
kubernetes==35.0.0
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.17
langchain-groq==1.1.2
langchain-huggingface==1.2.1
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
langgraph==1.0.10
langsmith==0.7.13
mcp==1.26.0
numpy==2.4.2
pydantic==2.12.5
pypdf==6.7.5
python-dotenv==1.2.2
regex==2026.2.28
SQLAlchemy==2.0.48
streamlit
sqlalchemy
